qc/seqlens: relocated to 'qc.apps.seqlens'.
